chore: adds interface file for icons.  translate some css into tailwind commands
